+++ hypothesis-python/docs/conf.py
@@ -33,6 +33,7 @@ extensions = [
     "sphinx.ext.extlinks",
     "sphinx.ext.viewcode",
     "sphinx.ext.intersphinx",
+    "sphinx.ext.napoleon",
     "hoverxref.extension",
     "sphinx_codeautolink",
     "sphinx_selective_exclude.eager_only",
@@ -45,6 +46,7 @@ extensions = [
 
 templates_path = ["_templates"]
 
+# config for hypothesis_redirects
 redirects = {
     "details": "reference/index.html",
     "data": "reference/strategies.html",
@@ -141,6 +143,7 @@ hoverxref_role_types = {
     "mod": "tooltip",
     "obj": "tooltip",
     "ref": "tooltip",
+    "data": "tooltip",
 }
 
 intersphinx_mapping = {
@@ -177,25 +180,38 @@ rst_prolog = """
 .. |note| replace:: :func:`~hypothesis.note`
 
 .. |max_examples| replace:: :obj:`~hypothesis.settings.max_examples`
-.. |settings.max_examples| replace:: :obj:`~hypothesis.settings.max_examples`
-.. |settings.database| replace:: :obj:`~hypothesis.settings.database`
-.. |settings.deadline| replace:: :obj:`~hypothesis.settings.deadline`
-.. |settings.derandomize| replace:: :obj:`~hypothesis.settings.derandomize`
-.. |settings.phases| replace:: :obj:`~hypothesis.settings.phases`
-.. |settings.print_blob| replace:: :obj:`~hypothesis.settings.print_blob`
-.. |settings.report_multiple_bugs| replace:: :obj:`~hypothesis.settings.report_multiple_bugs`
-.. |settings.verbosity| replace:: :obj:`~hypothesis.settings.verbosity`
-.. |settings.suppress_health_check| replace:: :obj:`~hypothesis.settings.suppress_health_check`
-.. |settings.stateful_step_count| replace:: :obj:`~hypothesis.settings.stateful_step_count`
-.. |settings.backend| replace:: :obj:`~hypothesis.settings.backend`
+.. |settings.max_examples| replace:: :obj:`settings.max_examples <hypothesis.settings.max_examples>`
+.. |settings.database| replace:: :obj:`settings.database <hypothesis.settings.database>`
+.. |settings.deadline| replace:: :obj:`settings.deadline <hypothesis.settings.deadline>`
+.. |settings.derandomize| replace:: :obj:`settings.derandomize <hypothesis.settings.derandomize>`
+.. |settings.phases| replace:: :obj:`settings.phases <hypothesis.settings.phases>`
+.. |settings.print_blob| replace:: :obj:`settings.print_blob <hypothesis.settings.print_blob>`
+.. |settings.report_multiple_bugs| replace:: :obj:`settings.report_multiple_bugs <hypothesis.settings.report_multiple_bugs>`
+.. |settings.verbosity| replace:: :obj:`settings.verbosity <hypothesis.settings.verbosity>`
+.. |settings.suppress_health_check| replace:: \
+    :obj:`settings.suppress_health_check <hypothesis.settings.suppress_health_check>`
+.. |settings.stateful_step_count| replace:: :obj:`settings.stateful_step_count <hypothesis.settings.stateful_step_count>`
+.. |settings.backend| replace:: :obj:`settings.backend <hypothesis.settings.backend>`
+
+.. |~settings.max_examples| replace:: :obj:`~hypothesis.settings.max_examples`
+.. |~settings.database| replace:: :obj:`~hypothesis.settings.database`
+.. |~settings.deadline| replace:: :obj:`~hypothesis.settings.deadline`
+.. |~settings.derandomize| replace:: :obj:`~hypothesis.settings.derandomize`
+.. |~settings.phases| replace:: :obj:`~hypothesis.settings.phases`
+.. |~settings.print_blob| replace:: :obj:`~hypothesis.settings.print_blob`
+.. |~settings.report_multiple_bugs| replace:: :obj:`~hypothesis.settings.report_multiple_bugs`
+.. |~settings.verbosity| replace:: :obj:`~hypothesis.settings.verbosity`
+.. |~settings.suppress_health_check| replace:: :obj:`~hypothesis.settings.suppress_health_check`
+.. |~settings.stateful_step_count| replace:: :obj:`~hypothesis.settings.stateful_step_count`
+.. |~settings.backend| replace:: :obj:`~hypothesis.settings.backend`
 
 .. |HealthCheck.data_too_large| replace:: :obj:`HealthCheck.data_too_large <hypothesis.HealthCheck.data_too_large>`
 .. |HealthCheck.filter_too_much| replace:: :obj:`HealthCheck.filter_too_much <hypothesis.HealthCheck.filter_too_much>`
 .. |HealthCheck.too_slow| replace:: :obj:`HealthCheck.too_slow <hypothesis.HealthCheck.too_slow>`
-.. |HealthCheck.function_scoped_fixture| replace:: :obj:`HealthCheck.function_scoped_fixture \
-<hypothesis.HealthCheck.function_scoped_fixture>`
-.. |HealthCheck.differing_executors| replace:: :obj:`HealthCheck.differing_executors \
-<hypothesis.HealthCheck.differing_executors>`
+.. |HealthCheck.function_scoped_fixture| replace:: \
+    :obj:`HealthCheck.function_scoped_fixture <hypothesis.HealthCheck.function_scoped_fixture>`
+.. |HealthCheck.differing_executors| replace:: \
+    :obj:`HealthCheck.differing_executors <hypothesis.HealthCheck.differing_executors>`
 .. |HealthCheck| replace:: :obj:`~hypothesis.HealthCheck`
 
 .. |Phase| replace:: :obj:`Phase <hypothesis.Phase>`
@@ -213,7 +229,7 @@ rst_prolog = """
 .. |Verbosity.quiet| replace:: :obj:`Verbosity.quiet <hypothesis.Verbosity.quiet>`
 
 .. |InvalidArgument| replace:: :obj:`InvalidArgument <hypothesis.errors.InvalidArgument>`
-
+.. |DidNotReproduce| replace:: :obj:`DidNotReproduce <hypothesis.errors.DidNotReproduce>`
 
 .. |st.lists| replace:: :func:`~hypothesis.strategies.lists`
 .. |st.integers| replace:: :func:`~hypothesis.strategies.integers`
@@ -259,9 +275,28 @@ rst_prolog = """
 .. |.map()| replace:: :func:`.map() <hypothesis.strategies.SearchStrategy.map>`
 .. |.example()| replace:: :func:`.example() <hypothesis.strategies.SearchStrategy.example>`
 
+.. |PrimitiveProvider| replace:: :class:`~hypothesis.internal.conjecture.providers.PrimitiveProvider`
+.. |PrimitiveProvider.realize| replace:: :func:`~hypothesis.internal.conjecture.providers.PrimitiveProvider.realize`
+.. |PrimitiveProvider.draw_integer| replace:: \
+    :func:`~hypothesis.internal.conjecture.providers.PrimitiveProvider.draw_integer`
+.. |PrimitiveProvider.draw_boolean| replace:: \
+    :func:`~hypothesis.internal.conjecture.providers.PrimitiveProvider.draw_boolean`
+.. |PrimitiveProvider.draw_float| replace:: :func:`~hypothesis.internal.conjecture.providers.PrimitiveProvider.draw_float`
+.. |PrimitiveProvider.draw_string| replace:: :func:`~hypothesis.internal.conjecture.providers.PrimitiveProvider.draw_string`
+.. |PrimitiveProvider.draw_bytes| replace:: :func:`~hypothesis.internal.conjecture.providers.PrimitiveProvider.draw_bytes`
+
+.. |AVAILABLE_PROVIDERS| replace:: :data:`~hypothesis.internal.conjecture.providers.AVAILABLE_PROVIDERS`
+.. |TESTCASE_CALLBACKS| replace:: :data:`~hypothesis.internal.observability.TESTCASE_CALLBACKS`
+.. |BUFFER_SIZE| replace:: :data:`~hypothesis.internal.conjecture.engine.BUFFER_SIZE`
+.. |MAX_SHRINKS| replace:: :data:`~hypothesis.internal.conjecture.engine.MAX_SHRINKS`
+.. |MAX_SHRINKING_SECONDS| replace:: :data:`~hypothesis.internal.conjecture.engine.MAX_SHRINKING_SECONDS`
+.. |BackendCannotProceed| replace:: :exc:`~hypothesis.errors.BackendCannotProceed`
+
 .. |@rule| replace:: :func:`@rule <hypothesis.stateful.rule>`
+.. |RuleBasedStateMachine| replace:: :class:`~hypothesis.stateful.RuleBasedStateMachine`
 
 .. |@reproduce_failure| replace:: :func:`@reproduce_failure <hypothesis.reproduce_failure>`
+.. |@seed| replace:: :func:`@seed <hypothesis.seed>`
 
 .. |ExampleDatabase| replace:: :class:`~hypothesis.database.ExampleDatabase`
 .. |ExampleDatabase.save| replace:: :func:`~hypothesis.database.ExampleDatabase.save`