+++ hypothesis-python/src/hypothesis/internal/conjecture/providers.py
@@ -37,6 +37,7 @@ from hypothesis.internal.cache import LRUCache
 from hypothesis.internal.compat import WINDOWS, int_from_bytes
 from hypothesis.internal.conjecture.choice import (
     ChoiceConstraintsT,
+    ChoiceT,
     ChoiceTypeT,
     FloatConstraints,
     choice_constraints_key,
@@ -63,6 +64,7 @@ from hypothesis.internal.floats import (
     next_up,
 )
 from hypothesis.internal.intervalsets import IntervalSet
+from hypothesis.internal.observability import InfoObservationType
 
 if TYPE_CHECKING:
     from typing import TypeAlias
@@ -75,16 +77,36 @@ LifetimeT: "TypeAlias" = Literal["test_case", "test_function"]
 COLLECTION_DEFAULT_MAX_SIZE = 10**10  # "arbitrarily large"
 
 
-# The available `PrimitiveProvider`s, and therefore also the available backends
-# for use by @settings(backend=...). The key is the name to be used in the backend=
-# value, and the value is the importable path to a subclass of PrimitiveProvider.
-#
-# See also
-# https://hypothesis.readthedocs.io/en/latest/strategies.html#alternative-backends-for-hypothesis.
-#
-# NOTE: the PrimitiveProvider interface is not yet stable. We may continue to
-# make breaking changes to it. (but if you want to experiment and don't mind
-# breakage, here you go!)
+#: Registered Hypothesis backends. This is a dictionary whose keys are the name
+#: to be used in |settings.backend|, and whose values are a string of the absolute
+#: importable path to a subclass of |PrimitiveProvider|, which Hypothesis will
+#: instantiate when your backend is requested by a test's |settings.backend| value.
+#:
+#: For example, the default Hypothesis backend is registered as:
+#:
+#: .. code-block:: python
+#:
+#:    from hypothesis.internal.conjecture.providers import AVAILABLE_PROVIDERS
+#:
+#:    AVAILABLE_PROVIDERS["hypothesis"] = "hypothesis.internal.conjecture.providers.HypothesisProvider"
+#:
+#: And can be used with:
+#:
+#: .. code-block:: python
+#:
+#:     from hypothesis import given, settings, strategies as st
+#:
+#:     @given(st.integers())
+#:     @settings(backend="hypothesis")
+#:     def f(n):
+#:         pass
+#:
+#: Though, as ``backend="hypothesis"`` is the default setting, the above would
+#: typically not have any effect.
+#:
+#: The purpose of mapping to an absolute importable path, rather than the actual
+#: |PrimitiveProvider| class, is to avoid slowing down Hypothesis startup times
+#: by only importing alternative backends when required.
 AVAILABLE_PROVIDERS = {
     "hypothesis": "hypothesis.internal.conjecture.providers.HypothesisProvider",
     "hypothesis-urandom": "hypothesis.internal.conjecture.providers.URandomProvider",
@@ -278,9 +300,12 @@ def _get_local_constants() -> Constants:
 
 
 class _BackendInfoMsg(TypedDict):
-    type: str
+    type: InfoObservationType
     title: str
     content: Union[str, dict[str, Any]]
+
+
+# TODO_DOCS: link to choice sequence explanation page
 
 
 class PrimitiveProvider(abc.ABC):
@@ -371,6 +396,20 @@ class PrimitiveProvider(abc.ABC):
         self,
         p: float = 0.5,
     ) -> bool:
+        """
+        Draw a boolean choice.
+
+        Parameters
+        ----------
+        p: float
+            The probability of returning ``True``. Between 0 and 1 inclusive.
+
+            Except for ``0`` and ``1``, the value of ``p`` is a hint provided by
+            Hypothesis, and may be ignored by the backend.
+
+            If ``0``, the provider must return ``False``. If ``1``, the provider
+            must return ``True``.
+        """
         raise NotImplementedError
 
     @abc.abstractmethod
@@ -379,10 +418,27 @@ class PrimitiveProvider(abc.ABC):
         min_value: Optional[int] = None,
         max_value: Optional[int] = None,
         *,
-        # weights are for choosing an element index from a bounded range
         weights: Optional[dict[int, float]] = None,
         shrink_towards: int = 0,
     ) -> int:
+        """
+        Draw an integer choice.
+
+        Parameters
+        ----------
+        min_value : int | None
+            (Inclusive) lower bound on the integer value. If ``None``, there is
+            no lower bound.
+        max_value : int | None
+            (Inclusive) upper bound on the integer value. If ``None``, there is
+            no upper bound.
+        weights: dict[int, float] | None
+            Maps keys in the range [``min_value``, ``max_value``] to the probability
+            of returning that key.
+        shrink_towards: int
+            The integer to shrink towards. This is not used during generation and
+            can be ignored by backends.
+        """
         raise NotImplementedError
 
     @abc.abstractmethod
@@ -394,6 +450,21 @@ class PrimitiveProvider(abc.ABC):
         allow_nan: bool = True,
         smallest_nonzero_magnitude: float,
     ) -> float:
+        """
+        Draw a float choice.
+
+        Parameters
+        ----------
+        min_value : float
+            (Inclusive) lower bound on the float value.
+        max_value : float
+            (Inclusive) upper bound on the float value.
+        allow_nan : bool
+            If ``False``, it is invalid to return ``math.nan``.
+        smallest_nonzero_magnitude : float
+            The smallest allowed nonzero magnitude. ``draw_float`` should not
+            return a float ``f`` if ``abs(f) < smallest_nonzero_magnitude``.
+        """
         raise NotImplementedError
 
     @abc.abstractmethod
@@ -404,6 +475,18 @@ class PrimitiveProvider(abc.ABC):
         min_size: int = 0,
         max_size: int = COLLECTION_DEFAULT_MAX_SIZE,
     ) -> str:
+        """
+        Draw a string choice.
+
+        Parameters
+        ----------
+        intervals : IntervalSet
+            The set of codepoints to sample from.
+        min_size : int
+            (Inclusive) lower bound on the string length.
+        max_size : int
+            (Inclusive) upper bound on the string length.
+        """
         raise NotImplementedError
 
     @abc.abstractmethod
@@ -412,27 +495,106 @@ class PrimitiveProvider(abc.ABC):
         min_size: int = 0,
         max_size: int = COLLECTION_DEFAULT_MAX_SIZE,
     ) -> bytes:
+        """
+        Draw a bytes choice.
+
+        Parameters
+        ----------
+        min_size : int
+            (Inclusive) lower bound on the bytes length.
+        max_size : int
+            (Inclusive) upper bound on the bytes length.
+        """
         raise NotImplementedError
 
+    def per_test_case_context_manager(self):
+        """
+        Returns a context manager which will be entered each time Hypothesis
+        starts generating and executing one test case, and exited when that test
+        case finishes generating and executing, including if any exception is
+        thrown.
+
+        In the lifecycle of a Hypothesis test, this is called before
+        generating strategy values for each test case. This is just before any
+        :ref:`custom executor <custom-function-execution>` is called.
+
+        Even if not returning a custom context manager, |PrimitiveProvider|
+        subclasses are welcome to override this method to know when Hypothesis
+        starts and ends the execution of a single test case.
+        """
+        return contextlib.nullcontext()
+
+    def realize(self, value: T, *, for_failure: bool = False) -> T:
+        """
+        Called whenever hypothesis requires a concrete (non-symbolic) value from
+        a potentially symbolic value. Hypothesis will not check that ``value`` is
+        symbolic before calling ``realize``, so you should handle the case where
+        ``value`` is non-symbolic.
+
+        The returned value should be non-symbolic.  If you cannot provide a value,
+        raise |BackendCannotProceed| with a value of ``"discard_test_case"``.
+
+        If ``for_failure`` is ``True``, the value is associated with a failing example.
+        In this case, the backend should spend substantially more effort when
+        attempting to realize the value, since it is important to avoid discarding
+        failing examples. Backends may still raise |BackendCannotProceed| when
+        ``for_failure`` is ``True``, if realization is truly impossible or if
+        realization takes significantly longer than expected (say, 5 minutes).
+        """
+        return value
+
+    def replay_choices(self, choices: tuple[ChoiceT, ...]) -> None:
+        """
+        Called when Hypothesis has discovered a choice sequence which the provider
+        may wish to enqueue to replay under its own instrumentation when we next
+        ask to generate a test case, rather than generating one from scratch.
+
+        This is used to e.g. warm-start :pypi:`hypothesis-crosshair` with a corpus
+        of high-code-coverage inputs discovered by
+        `HypoFuzz <https://hypofuzz.com/>`_.
+        """
+        return None
+
+    def observe_test_case(self) -> dict[str, Any]:
+        """Called at the end of the test case when :ref:`observability
+        <observability>` is enabled.
+
+        The return value should be a non-symbolic json-encodable dictionary,
+        and will be included in observations as ``observation["metadata"]["backend"]``.
+        """
+        return {}
+
+    def observe_information_messages(
+        self, *, lifetime: _Lifetime
+    ) -> Iterable[_BackendInfoMsg]:
+        """Called at the end of each test case and again at end of the test function.
+
+        Return an iterable of ``{type: info/alert/error, title: str, content: str | dict}``
+        dictionaries to be delivered as individual information messages. Hypothesis
+        adds the ``run_start`` timestamp and ``property`` name for you.
+        """
+        assert lifetime in ("test_case", "test_function")
+        yield from []
+
     def span_start(self, label: int, /) -> None:  # noqa: B027  # non-abstract noop
-        """Marks the beginning of a semantically meaningful span.
+        """Marks the beginning of a semantically meaningful span of choices.
 
         Providers can optionally track this data to learn which sub-sequences
         of draws correspond to a higher-level object, recovering the parse tree.
-        `label` is an opaque integer, which will be shared by all spans drawn
+        ``label`` is an opaque integer, which will be shared by all spans drawn
         from a particular strategy.
 
-        This method is called from ConjectureData.start_span().
+        This method is called from ``ConjectureData.start_span()``.
         """
 
     def span_end(self, discard: bool, /) -> None:  # noqa: B027, FBT001
-        """Marks the end of a semantically meaningful span.
+        """Marks the end of a semantically meaningful span of choices.
 
-        `discard` is True when the draw was filtered out or otherwise marked as
-        unlikely to contribute to the input data as seen by the user's test.
+        ``discard`` is ``True`` when the draw was filtered out or otherwise marked
+        as unlikely to contribute to the input data as seen by the user's test.
         Note however that side effects can make this determination unsound.
 
-        This method is called from ConjectureData.stop_span().
+        This method is called from ``ConjectureData.stop_span()``.
         """